Add support for windows as a client
