+++ rkvm-input/Cargo.toml
@@ -10,11 +10,18 @@ edition = "2021"
 serde = { version = "1.0.117", features = ["derive"] }
 futures = "0.3.8"
 smallvec = { version = "1.10.0", features = ["serde"] }
-inotify = "0.10.0"
 tokio = { version = "1.0.1", features = ["fs", "io-util", "net", "sync", "rt", "time", "macros"] }
-libc = "0.2.77"
 thiserror = "1.0.40"
 tracing = "0.1.37"
+
+[target.'cfg(windows)'.dependencies]
+windows = { version = "0.52", features = [
+    "Win32_UI_Input_KeyboardAndMouse",
+    "Win32_Foundation"
+] }
+[target.'cfg(unix)'.dependencies]
+inotify = "0.10.0"
+libc = "0.2.77"
 
 [build-dependencies]
 bindgen = "0.65.1"
