+++ rkvm-input/src/linux/glue.rs
@@ -1,3 +1,2 @@
 #![allow(warnings)]
-
 include!(concat!(env!("OUT_DIR"), "/glue.rs"));
