+++ rkvm-input/src/linux/registry.rs
